refactor: centralize config literals and relocate fs security tests
