+++ src/notewise/_constants.py
@@ -10,6 +10,16 @@ CACHE_DB_FILENAME = ".notewise_cache.db"
 STATE_DIR_NAME = ".notewise"
 LOGS_DIR_NAME = "logs"
 CONFIG_FILENAME = "config.env"
+CONFIG_TEMP_SUFFIX = ".tmp"
+CONFIG_FILE_HEADER_LINES = (
+    "# notewise Configuration\n",
+    "# Generated by notewise setup wizard\n\n",
+)
+CONFIG_PRIORITY_KEY_ORDER = (
+    "DEFAULT_MODEL",
+    "OUTPUT_DIR",
+    "MAX_CONCURRENT_VIDEOS",
+)
 ALLOW_UNLISTED_MODELS_CONFIG_KEY = "ALLOW_UNLISTED_MODELS"
 ALLOW_UNLISTED_MODELS_ATTR = "allow_unlisted_models"
 OUTPUT_DIR_CONFIG_KEY = "OUTPUT_DIR"
