+++ src/notewise/ui/setup_wizard.py
@@ -10,8 +10,11 @@ from typing import TYPE_CHECKING, Any, Protocol
 from notewise._constants import (
     AUTH_TYPE_API_KEY,
     AUTH_TYPE_OAUTH_DEVICE,
+    CONFIG_FILE_HEADER_LINES,
     CONFIG_FILE_PERMISSION_MODE,
     CONFIG_FILENAME,
+    CONFIG_PRIORITY_KEY_ORDER,
+    CONFIG_TEMP_SUFFIX,
     DEFAULT_MAX_CONCURRENT_VIDEOS,
     DEFAULT_OUTPUT_DIR,
     OAUTH_SETUP_RUN_PROMPT,
@@ -129,7 +132,9 @@ def save_config(
     # No world-readable window exists because the temp file is restricted
     # before any content is written.
     temp_fd, temp_name = tempfile.mkstemp(
-        dir=config_path.parent, prefix=f"{CONFIG_FILENAME}.", suffix=".tmp"
+        dir=config_path.parent,
+        prefix=f"{CONFIG_FILENAME}.",
+        suffix=CONFIG_TEMP_SUFFIX,
     )
     temp_path = Path(temp_name)
     try:
@@ -138,20 +143,15 @@ def save_config(
             fchmod(temp_fd, CONFIG_FILE_PERMISSION_MODE)
         with os.fdopen(temp_fd, "w", encoding="utf-8") as f:
             temp_fd = -1  # Ownership transferred to the file object.
-            f.write("# notewise Configuration\n")
-            f.write("# Generated by notewise setup wizard\n\n")
+            for header_line in CONFIG_FILE_HEADER_LINES:
+                f.write(header_line)
 
-            priority_keys = [
-                "DEFAULT_MODEL",
-                "OUTPUT_DIR",
-                "MAX_CONCURRENT_VIDEOS",
-            ]
-            for key in priority_keys:
+            for key in CONFIG_PRIORITY_KEY_ORDER:
                 if key in current_config:
                     f.write(f"{key}={current_config[key]}\n")
 
             for key, value in sorted(current_config.items()):
-                if key not in priority_keys:
+                if key not in CONFIG_PRIORITY_KEY_ORDER:
                     f.write(f"{key}={value}\n")
 
         Path(temp_path).replace(config_path)
