+++ tests/unit/ui/test_setup_wizard.py
@@ -5,7 +5,11 @@ from unittest.mock import MagicMock, mock_open, patch
 
 import pytest
 
-from notewise._constants import CONFIG_FILE_PERMISSION_MODE, CONFIG_FILENAME
+from notewise._constants import (
+    CONFIG_FILE_PERMISSION_MODE,
+    CONFIG_FILENAME,
+    CONFIG_TEMP_SUFFIX,
+)
 from notewise.errors import ConfigurationError
 from notewise.ui.setup_wizard import (
     get_api_key,
@@ -119,75 +123,71 @@ class TestConfigIO:
         with pytest.raises(RuntimeError, match="bug"):
             load_config()
 
-    def test_save_config(self):
+    def test_save_config(self, mocker):
         """Test saving configuration merges with existing."""
         from pathlib import Path
 
         mock_path = Path("dummy_path")
-        with (
-            patch(
-                "notewise.ui.setup_wizard.load_config",
-                return_value={"OLD_KEY": "old_val"},
-            ),
-            patch(
-                "notewise.ui.setup_wizard.tempfile.mkstemp",
-                return_value=(3, "dummy_path.tmp"),
-            ) as mock_mkstemp,
-            patch("notewise.ui.setup_wizard.os.fchmod", create=True) as mock_fchmod,
-            patch("notewise.ui.setup_wizard.os.fdopen", mock_open()) as mock_file,
-            patch("notewise.ui.setup_wizard.os.replace") as mock_replace,
-            patch("notewise.ui.setup_wizard.get_config_path", return_value=mock_path),
-        ):
-            new_config = {"NEW_KEY": "new_val", "DEFAULT_MODEL": "new_model"}
-            save_config(new_config)
+        mocker.patch(
+            "notewise.ui.setup_wizard.load_config",
+            return_value={"OLD_KEY": "old_val"},
+        )
+        mock_mkstemp = mocker.patch(
+            "notewise.ui.setup_wizard.tempfile.mkstemp",
+            return_value=(3, "dummy_path.tmp"),
+        )
+        mock_fchmod = mocker.patch("notewise.ui.setup_wizard.os.fchmod", create=True)
+        mock_file = mocker.patch("notewise.ui.setup_wizard.os.fdopen", mock_open())
+        mock_replace = mocker.patch("notewise.ui.setup_wizard.os.replace")
+        mocker.patch("notewise.ui.setup_wizard.get_config_path", return_value=mock_path)
 
-            # Verify file write operations
-            handle = mock_file()
-            written_content = "".join(
-                call.args[0] for call in handle.write.call_args_list
-            )
+        new_config = {"NEW_KEY": "new_val", "DEFAULT_MODEL": "new_model"}
+        save_config(new_config)
 
-            assert "OLD_KEY=old_val" in written_content
-            assert "NEW_KEY=new_val" in written_content
-            assert "DEFAULT_MODEL=new_model" in written_content
-            mock_mkstemp.assert_called_once_with(
-                dir=mock_path.parent,
-                prefix=f"{CONFIG_FILENAME}.",
-                suffix=".tmp",
-            )
-            mock_fchmod.assert_called_once_with(3, CONFIG_FILE_PERMISSION_MODE)
-            mock_replace.assert_called_once_with(Path("dummy_path.tmp"), mock_path)
+        # Verify file write operations
+        handle = mock_file()
+        written_content = "".join(call.args[0] for call in handle.write.call_args_list)
 
-    def test_save_config_strips_legacy_youtube_auth_keys(self):
+        assert "OLD_KEY=old_val" in written_content
+        assert "NEW_KEY=new_val" in written_content
+        assert "DEFAULT_MODEL=new_model" in written_content
+        mock_mkstemp.assert_called_once_with(
+            dir=mock_path.parent,
+            prefix=f"{CONFIG_FILENAME}.",
+            suffix=CONFIG_TEMP_SUFFIX,
+        )
+        mock_fchmod.assert_called_once_with(3, CONFIG_FILE_PERMISSION_MODE)
+        mock_replace.assert_called_once_with(Path("dummy_path.tmp"), mock_path)
+
+    def test_save_config_strips_legacy_youtube_auth_keys(self, mocker):
         """Saving config should remove legacy OAuth and cookie-era auth keys."""
         from pathlib import Path
 
         mock_path = Path("dummy_path")
-        with (
-            patch(
-                "notewise.ui.setup_wizard.load_config",
-                return_value={
-                    "YOUTUBE_USE_OAUTH": "true",
-                    "YOUTUBE_SAVE_OAUTH_TOKEN": "true",
-                    "YOUTUBE_OAUTH_TOKEN_FILE": "/tmp/token.json",
-                    "YOUTUBE_AUTO_REFRESH_OAUTH_TOKEN": "false",
-                    "OLD_KEY": "old_val",
-                },
-            ),
-            patch(
-                "notewise.ui.setup_wizard.tempfile.mkstemp",
-                return_value=(3, "dummy_path.tmp"),
-            ),
-            patch("notewise.ui.setup_wizard.os.fchmod", create=True),
-            patch("notewise.ui.setup_wizard.os.fdopen", mock_open()) as mock_file,
-            patch("notewise.ui.setup_wizard.os.replace"),
-            patch("notewise.ui.setup_wizard.get_config_path", return_value=mock_path),
-        ):
-            save_config({"DEFAULT_MODEL": "new_model"})
+        mocker.patch(
+            "notewise.ui.setup_wizard.load_config",
+            return_value={
+                "YOUTUBE_USE_OAUTH": "true",
+                "YOUTUBE_SAVE_OAUTH_TOKEN": "true",
+                "YOUTUBE_OAUTH_TOKEN_FILE": "/tmp/token.json",
+                "YOUTUBE_AUTO_REFRESH_OAUTH_TOKEN": "false",
+                "OLD_KEY": "old_val",
+            },
+        )
+        mocker.patch(
+            "notewise.ui.setup_wizard.tempfile.mkstemp",
+            return_value=(3, "dummy_path.tmp"),
+        )
+        mocker.patch("notewise.ui.setup_wizard.os.fchmod", create=True)
+        mock_file = mocker.patch("notewise.ui.setup_wizard.os.fdopen", mock_open())
+        mocker.patch("notewise.ui.setup_wizard.os.replace")
+        mocker.patch("notewise.ui.setup_wizard.get_config_path", return_value=mock_path)
 
-            written_content = "".join(
-                call.args[0] for call in mock_file().write.call_args_list
-            )
+        save_config({"DEFAULT_MODEL": "new_model"})
+
+        written_content = "".join(
+            call.args[0] for call in mock_file().write.call_args_list
+        )
 
         assert "OLD_KEY=old_val" in written_content
         assert "DEFAULT_MODEL=new_model" in written_content
@@ -222,27 +222,6 @@ class TestConfigIO:
         config_path = get_config_path()
         mode = stat.S_IMODE(config_path.stat().st_mode)
         assert mode == 0o600
-
-    def test_save_config_does_not_truncate_symlink_target(self, tmp_path, monkeypatch):
-        """save_config must swap a planted symlink, not truncate its victim."""
-        monkeypatch.setenv("NOTEWISE_HOME", str(tmp_path / "state"))
-        monkeypatch.delenv("DEFAULT_MODEL", raising=False)
-
-        victim = tmp_path / "victim.env"
-        victim.write_text("VICTIM=keep-me\n", encoding="utf-8")
-        config_path = get_config_path()
-        try:
-            config_path.symlink_to(victim)
-        except OSError:
-            pytest.skip("symlink creation is not available on this platform")
-
-        save_config({"DEFAULT_MODEL": "gemini/gemini-2.5-flash"})
-
-        assert victim.read_text(encoding="utf-8") == "VICTIM=keep-me\n"
-        assert "DEFAULT_MODEL=gemini/gemini-2.5-flash" in config_path.read_text(
-            encoding="utf-8"
-        )
-        assert config_path.is_symlink() is False
 
     def test_show_current_config_masks_api_keys(self):
         """Read-only config display should mask secret values."""
